refactoring and it works for whitelabel now
